Metrics "Heat" map   - First iteration of heap map implemented.     - Links now change color based on a few factors. Mainly (how many times that link occurred) / (number of total links / 2) with a few other things thrown in. BUG: The Heat Map does not work if there was a link recorded coming from / going to a Building Site. TODO: Fix that bug.

+++ Assets/Scripts/Database Scripts/Metrics/MetricDisplay.js
@@ -22,8 +22,8 @@ public class MetricDisplay
 	private var averageTimesUndone : int = 0;	
 	private var undoCount : int = 0;
 	
-	public var linkArray : int[,];	
-	public var numberOfLinks : int = 0;	
+	public var linkArray : float[,];	
+	public var numberOfLinks : float = 0;	
 	
 	public function MetricDisplay()
 	{
@@ -138,7 +138,7 @@ public class MetricDisplay
 	
 	public function CreateLinkArray(numberOfBuildings : int)
 	{
-		linkArray = new int[numberOfBuildings, numberOfBuildings];
+		linkArray = new float[numberOfBuildings, numberOfBuildings];
 		
 		for(var i = 0; i < LinkDataList.Count; i++)
 		{
